Refactor Modal component.
