+++ assets/js/components/permissions-modal/index.js
@@ -28,7 +28,7 @@ import { useEffect, useCallback } from '@wordpress/element';
 import Data from 'googlesitekit-data';
 import { STORE_NAME as CORE_USER } from '../../googlesitekit/datastore/user/constants';
 import Dialog from '../dialog';
-import Modal from '../modal';
+import Modal from '../Modal';
 import { snapshotAllStores } from '../../googlesitekit/data/create-snapshot-store';
 
 const { useSelect, useDispatch, useRegistry } = Data;
