+++ assets/js/components/user-menu.js
@@ -30,7 +30,7 @@ import { getSiteKitAdminURL, clearWebStorage } from '../util';
 import Dialog from './dialog';
 import Button from './button';
 import Menu from './menu';
-import Modal from './modal';
+import Modal from './Modal';
 import { STORE_NAME as CORE_SITE } from '../googlesitekit/datastore/site/constants';
 import { STORE_NAME as CORE_USER } from '../googlesitekit/datastore/user/constants';
 const { withSelect } = Data;
